Moving package, removing exception class
